requirements.txt
aiogram==3.22.0
apscheduler==3.11.0
openai==2.0.1
python-dotenv==1.1.1
requests==2.32.5